from int to double

+++ src/Main.java
@@ -1,12 +1,12 @@
 public class Main{
     public static void main(String[] args) {
-        double a;
+        int a;
         int b;
         a = 10;
         b = 25;
         System.out.println(a + b);
         System.out.println(a - b);
-        System.out.println(a / b);
+        System.out.println((double)a / b);
         System.out.println(a * b);
         System.out.println(a % b);
         System.out.println("a+ b + \"Jaa\" + 4");
